ajuste inicial componentes
