--- src/components/MockButton/styles.ts
@@ -1,22 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  container: {
-    backgroundColor: 'transparent',
-    borderWidth: 2,
-    borderColor: '#F59E0B',
-    borderStyle: 'dashed',
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-    borderRadius: 8,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginVertical: 8,
-    width: '100%',
-  },
-  text: {
-    color: '#F59E0B',
-    fontSize: 14,
-    fontWeight: 'bold',
-  }
-});
