--- src/components/ReturnHomeArrow/ReturnHomeArrowViewModel.ts
@@ -1,14 +1,0 @@
-import { useNavigation, NavigationProp } from '@react-navigation/native';
-import { RootStackParamList } from '../../../App';
-
-export function useReturnHomeArrowViewModel() {
-    const navigation = useNavigation<NavigationProp<RootStackParamList>>();
-
-    const handlePress = () => {
-        navigation.navigate("Home");
-    };
-
-    return {
-        handlePress,
-    };
-};
