--- src/components/ReturnHomeArrow/styles.ts
@@ -1,10 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    returnArrowContainer: {
-        flexDirection: "row",
-        alignItems: "center",
-        paddingHorizontal: 16,
-        paddingVertical: 12,
-    }
-});
