+++ src/screens/PerfilSorteado/components/PerfilSorteadoHeader/PerfilSorteadoHeaderViewModel.test.ts
@@ -1,9 +1,16 @@
 import { renderHook } from '@testing-library/react-native';
 import { usePerfilSorteadoHeaderViewModel } from './PerfilSorteadoHeaderViewModel';
-import { calcularIdade } from '../../utils/Usuario/calcularIdade';
+import { calcularIdade } from '../../../../utils/Usuario/calcularIdade';
 
-jest.mock('../../utils/Usuario/calcularIdade', () => ({
+jest.mock('../../../../utils/Usuario/calcularIdade', () => ({
   calcularIdade: jest.fn(),
+}));
+
+const mockNavigate = jest.fn();
+jest.mock('@react-navigation/native', () => ({
+  useNavigation: () => ({
+    navigate: mockNavigate,
+  }),
 }));
 
 describe('usePerfilSorteadoHeaderViewModel', () => {
@@ -36,5 +43,21 @@ describe('usePerfilSorteadoHeaderViewModel', () => {
 
     expect(calcularIdade).toHaveBeenCalledWith('1990-01-01');
     expect(result.current.idade).toBe(30);
+  });
+
+  it('deve chamar navigation.navigate com "Home" quando handleReturnHome for chamado', () => {
+    (calcularIdade as jest.Mock).mockReturnValue(25);
+
+    const { result } = renderHook<
+      ReturnType<typeof usePerfilSorteadoHeaderViewModel>,
+      Props
+    >(
+      (props) => usePerfilSorteadoHeaderViewModel(props),
+      { initialProps: mockProps }
+    );
+
+    result.current.handleReturnHome();
+
+    expect(mockNavigate).toHaveBeenCalledWith('Home');
   });
 });
