+++ src/screens/PerfilSorteado/components/PerfilSorteadoHeader/styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { theme } from "../../styles/theme";
+import { theme } from "../../../../styles/theme";
 
 export const styles = StyleSheet.create({
     headerContainer: {
