+++ src/screens/PerfilSorteado/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { View, ScrollView } from 'react-native';
 
-import { PerfilSorteadoHeader } from '../../components/PerfilSorteadoHeader';
-import { PerfilSorteadoContent } from '../../components/PerfilSorteadoContent';
+import { PerfilSorteadoHeader } from './components/PerfilSorteadoHeader';
+import { PerfilSorteadoContent } from './components/PerfilSorteadoContent';
 import { usePerfilSorteadoViewModel } from './PerfilSorteadoViewModel';
 import { styles } from './styles';
